FCT ajout DELETE Nation(contraintes), modifier service (2en1 requête)
- Ajout la méthode deleteNation  aux services deleteNation
- Ajout contraintes dans Net.Core dans l'action DELETE (ne pas supprimer le parent si des enfants existent)
- Ajout getStrategicRoleById/service
- REFACTOR (getAllTanks, getListOfTanksByType, getListOfTanksByNation) GET Sous-requêtes HTTP
- afficher le nom du rôle stratégique dans les tableaux
- Modal delete Nation

+++ TP_Tankopedia_ASP/TP_Tankopedia_ASP/Controllers/NationsController.cs
@@ -109,6 +109,11 @@ namespace TP_Tankopedia_ASP.Controllers
             {
                 return NotFound();
             }
+            var associatedTanks = _context.Tanks.Where(x => x.Id == id).Any();
+            if (associatedTanks)
+            {
+                return StatusCode(StatusCodes.Status403Forbidden, new { Message = "We can't wipe out the nation that has tanks " });
+            }
 
             _context.Nations.Remove(nation);
             await _context.SaveChangesAsync();
